requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
requests==2.31.0
dj-database-url==2.2.0
# CV grading ML
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.11.0
nltk>=3.8.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
python-docx>=1.0.0